+++ static/client/components/Views/TableView/TableView.tsx
@@ -1,56 +1,22 @@
 import React, { useMemo } from "react";
 
-import { Accordion, Badge, Tooltip } from "@canonical/react-components";
+import { Accordion, Tooltip } from "@canonical/react-components";
 
-import TableViewRow from "./TableViewRow";
+import ProjectContent from "./ProjectContent";
+import ProjectTitle from "./ProjectTitle";
 
-import TableViewRowItem from "@/components/Views/TableView/TableViewRowItem";
 import { useProjects } from "@/services/api/hooks/projects";
-import type { IPage, IPagesResponse } from "@/services/api/types/pages";
 
 const TableView: React.FC = () => {
   const { data: projects } = useProjects();
 
-  const getAccordionBadgeCount = useMemo(() => {
-    function countPages(page: IPage): number {
-      if (!page) return 0;
-      return 1 + page.children?.reduce((acc, child) => acc + countPages(child), 0);
-    }
-
-    return (project: IPagesResponse["data"]) => countPages(project.templates);
-  }, []);
-
-  const getAccordionTitle = useMemo(() => {
-    return (project: IPagesResponse["data"]) => (
-      <span style={{ display: "inline-flex", alignItems: "baseline" }}>
-        <span className="p-muted-heading" style={{ margin: "0 0.5rem" }}>
-          {project.name}
-        </span>
-        <Badge value={getAccordionBadgeCount(project)} />
-      </span>
-    );
-  }, [getAccordionBadgeCount]);
-
-  const getAccordionContent = useMemo(() => {
-    return (project: IPagesResponse["data"]) => (
-      <table>
-        <tbody>
-          <TableViewRow page={project.templates} />
-          {project.templates.children.map((page) => (
-            <TableViewRowItem key={page.url} page={page} />
-          ))}
-        </tbody>
-      </table>
-    );
-  }, []);
-
   const getAccordionSections = useMemo(() => {
     if (!projects?.length) return [];
     return projects.map((project) => ({
-      title: getAccordionTitle(project),
-      content: getAccordionContent(project),
+      title: <ProjectTitle project={project} />,
+      content: <ProjectContent project={project} />,
     }));
-  }, [projects, getAccordionTitle, getAccordionContent]);
+  }, [projects]);
 
   return (
     <>
